+++ cmd/mailchain/internal/serve.go
@@ -20,7 +20,7 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/mailchain/mailchain/cmd/mailchain/config"
 	"github.com/mailchain/mailchain/cmd/mailchain/config/defaults"
-	"github.com/mailchain/mailchain/internal/pkg/http/rest/addresses"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/rest/handlers"
 	"github.com/mailchain/mailchain/internal/pkg/http/rest/ethereum/address/messages"
 	"github.com/mailchain/mailchain/internal/pkg/http/rest/ethereum/address/publickey"
 	"github.com/mailchain/mailchain/internal/pkg/http/rest/ethereum/messages/send"
@@ -74,7 +74,7 @@ func CreateRouter(cmd *cobra.Command) (http.Handler, error) {
 	deriveKeyOptions := multi.OptionsBuilders{
 		Scrypt: []scrypt.DeriveOptionsBuilder{scrypt.WithPassphrase(passphrase)},
 	}
-	r.HandleFunc("/api/addresses", addresses.GetAddresses(keystore)).Methods("GET")
+	r.HandleFunc("/api/addresses", handlers.GetAddresses(keystore)).Methods("GET")
 	r.HandleFunc("/api/ethereum/{network}/address/{address:[-0-9a-zA-Z]+}/public-key", publickey.GetPublicKey(pubKeyFinders)).Methods("GET")
 	r.HandleFunc(
 		"/api/ethereum/{network}/address/{address:[-0-9a-zA-Z]+}/messages",
